requirements.txt
pydantic>=2.0
pyyaml>=6.0
openai>=1.0
anthropic>=0.20
python-telegram-bot>=20.0
discord.py>=2.0
slack-sdk>=3.20
httpx>=0.24
